requirements.txt
Django==2.1.1
easy-thumbnails==2.7
django-filer==1.7.1
django-suit==0.2.26
django-robots==3.1.0
django-cleanup==2.1.0
django-admin-sortable2==0.7.5
django-admin-sortable==2.1.17
django-ace==1.0.7
django-ckeditor==5.6.1
django-webpack-loader==0.6.0
django-sitetree==1.12.0
django-environ==0.4.5
psycopg2==2.8.4
yturl==2.0.2
simplejson==3.16.0
lxml==4.4.0
